pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "modelswitch"
description = "LLM Gateway Proxy with multi-provider fallback"
readme = "README.md"
requires-python = ">=3.11"
dynamic = ["version"]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "openai>=1.30.0",
    "anthropic>=0.30.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "prometheus-client>=0.21.0",
    "watchdog>=4.0.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "coverage>=7.0",
    "pytest-timeout>=2.0",
]

[project.scripts]
modelswitch = "app.cli:main"

[tool.setuptools_scm]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.setuptools.package-data]
"app" = ["web/**/*", "config.yaml.example"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
